Taking online out of content and alt text
I actually want to monitor changes to our rankings for phrases containing online by taking it out of the alt text and then I'll put it back in the alt text, and then after its cached, I'll add it into the content. I'm doing a test for a blog post.